% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{folder_reference_factory}
\alias{folder_reference_factory}
\title{Create a folder reference function}
\usage{
folder_reference_factory(.target_folder)
}
\arguments{
\item{.target_folder}{The root folder path}
}
\value{
A function that builds paths relative to .target_folder
}
\description{
Factory function that creates a path builder for a specific target folder.
}
