% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{generate_search_hash}
\alias{generate_search_hash}
\title{Generate hash for search parameters}
\usage{
generate_search_hash(
  query,
  max_k = 100,
  min_year = NULL,
  journal_filter = NULL,
  journal_name = NULL,
  title_keyword = NULL,
  author_keyword = NULL
)
}
\arguments{
\item{query}{Text query}

\item{max_k}{Maximum results}

\item{min_year}{Minimum year}

\item{journal_filter}{Category filter}

\item{journal_name}{Journal name filter}

\item{title_keyword}{Title keyword filter}

\item{author_keyword}{Author keyword filter}
}
\value{
8-character hash string
}
\description{
Creates a consistent hash from search parameters using digest package.
}
