% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folders.R
\name{get_folder_refs}
\alias{get_folder_refs}
\title{Get folder reference functions}
\usage{
get_folder_refs(config = NULL)
}
\arguments{
\item{config}{Optional config list. If NULL, will load from get_folder_config()}
}
\value{
A list of folder reference functions
}
\description{
Returns a list of folder reference functions for easy path building.
}
