% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_last_updated}
\alias{get_last_updated}
\title{Returns the last recorded database rebuild or content update date.}
\usage{
get_last_updated(pool = NULL)
}
\arguments{
\item{pool}{Database pool. Defaults to get_api_pool()}
}
\value{
Date string in YYYY-MM-DD format, or NULL if not recorded
}
\description{
Returns the last recorded database rebuild or content update date.
}
