% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_processed_handles}
\alias{get_processed_handles}
\title{Get list of processed paper handles}
\usage{
get_processed_handles(con)
}
\arguments{
\item{con}{DuckDB connection}
}
\value{
Character vector of handles
}
\description{
Returns all paper handles already in the database.
}
