% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_saved_search}
\alias{get_saved_search}
\title{Get a saved search by hash}
\usage{
get_saved_search(hash, pool = NULL)
}
\arguments{
\item{hash}{8 character search hash}

\item{pool}{Database pool}
}
\value{
List or NULL if not found
}
\description{
Retrieves a saved search and its results.
}
