% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_total_articles}
\alias{get_total_articles}
\title{Get total article count}
\usage{
get_total_articles(pool = NULL)
}
\arguments{
\item{pool}{Database pool. Defaults to get_api_pool()}
}
\value{
Integer count
}
\description{
Returns the total number of articles in the database.
}
