% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_stats.R
\name{init_handle_stats_table}
\alias{init_handle_stats_table}
\title{Initialize handle_stats table}
\usage{
init_handle_stats_table(con)
}
\arguments{
\item{con}{DuckDB connection}
}
\value{
TRUE invisibly
}
\description{
Creates the handle_stats table schema if it doesn't exist.
}
