% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_all_journals}
\alias{parse_all_journals}
\title{Parse all journals that need updating}
\usage{
parse_all_journals(
  repec_folder = NULL,
  rds_folder = NULL,
  script_path = NULL,
  skip_today = TRUE
)
}
\arguments{
\item{repec_folder}{Path to RePEc folder. Defaults to config$repec_folder}

\item{rds_folder}{Path to RDS output folder. Defaults to config$rds_folder}

\item{script_path}{Path to Perl parsing script}

\item{skip_today}{If TRUE, skips repos already parsed today}
}
\value{
Data frame of files that were parsed invisibly
}
\description{
Main function to parse all ReDIF files that are new or updated.
}
