% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_and_save_redif}
\alias{parse_and_save_redif}
\title{Parse and save ReDIF files for a repository}
\usage{
parse_and_save_redif(
  repo_files,
  rds_folder = NULL,
  script_path = "parse_redif_simple.pl"
)
}
\arguments{
\item{repo_files}{Data frame with file paths for one repository}

\item{rds_folder}{Path to RDS output folder. Defaults to config$rds_folder}

\item{script_path}{Path to Perl parsing script}
}
\value{
Parsed data invisibly
}
\description{
Parses all ReDIF files for a given repository and saves result as RDS.
}
