% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{restore_db_from_parquet}
\alias{restore_db_from_parquet}
\title{Restore database from Parquet file}
\usage{
restore_db_from_parquet(pqt_file, db_path = NULL)
}
\arguments{
\item{pqt_file}{Path to parquet file}

\item{db_path}{Path to DuckDB database. Defaults to config$db_folder/articles.duckdb}
}
\value{
Path to database invisibly
}
\description{
Recreates the database from a Parquet backup file, including indices.
}
