% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_email.R
\name{send_email}
\alias{send_email}
\title{send_email}
\usage{
send_email(
  from,
  passwd,
  to = "<fr904103@bmrn.com>",
  subject = "test",
  body = "This was sent from R",
  smtpServer = "mail.bmrn.com"
)
}
\description{
send_email
}
