% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_repec_cpd_conf}
\alias{sync_repec_cpd_conf}
\title{Sync RePEc CPD conf papers}
\usage{
sync_repec_cpd_conf(
  dest_root = here::here("RePEc"),
  rsync_bin = "/opt/homebrew/bin/rsync"
)
}
\arguments{
\item{dest_root}{Root folder for RePEc data. Defaults to here::here("RePEc")}

\item{rsync_bin}{Path to rsync binary}
}
\value{
Path to synced folder (invisibly)
}
\description{
Downloads or updates related works data using rsync from rsync.repec.org.
}
