% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync.R
\name{sync_repec_pers}
\alias{sync_repec_pers}
\title{Sync RePEc author identification data (pers archive)}
\usage{
sync_repec_pers(dest_root = NULL, rsync_bin = "/opt/homebrew/bin/rsync")
}
\arguments{
\item{dest_root}{Root folder for RePEc data. Defaults to config$repec_folder}

\item{rsync_bin}{Path to rsync binary}
}
\value{
Path to synced pers folder (invisibly)
}
\description{
Downloads or updates the RePEc person (pers) archive, which contains
author identifiers and metadata in RDF format.
}
