% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{restore_db_from_parquet}
\alias{restore_db_from_parquet}
\title{Restore database from Parquet files}
\usage{
restore_db_from_parquet(pqt_folder = NULL, date_stamp = NULL, db_path = NULL)
}
\arguments{
\item{pqt_folder}{Path to folder containing parquet files}

\item{date_stamp}{Backup suffix. If NULL, picks the most recent found in articles_*.parquet}

\item{db_path}{Path to DuckDB database}
}
\value{
Path to database invisibly
}
\description{
Recreates the database from Parquet backup files, including indices.
}
